Added collision for entities. Switched from float to double.

+++ VoxEngine/Core/Shapes/Box.cpp
@@ -1,11 +1,11 @@
 #include "Box.h"
 
 Box::Box() :
-	center(0.0f), halfExtents(0.0f)
+	center(0.0), halfExtents(0.0)
 {
 }
 
-Box::Box(const glm::vec3& center, const glm::vec3& halfExtents) :
+Box::Box(const glm::dvec3& center, const glm::dvec3& halfExtents) :
 	center(center), halfExtents(halfExtents)
 {
 }
